Moved Window Class code from sdl2_sparkling.c to its own
